+++ dizoo/classic_control/cartpole/config/__init__.py
@@ -1,21 +1,21 @@
-from .cartpole_dqn_config import cartpole_dqn_config, cartpole_dqn_create_config
 from .cartpole_a2c_config import cartpole_a2c_config, cartpole_a2c_create_config
-from .cartpole_impala_config import cartpole_impala_config, cartpole_impala_create_config
-from .cartpole_ppo_config import cartpole_ppo_config, cartpole_ppo_create_config
-from .cartpole_ppo_offpolicy_config import cartpole_ppo_offpolicy_config, cartpole_ppo_offpolicy_create_config
-from .cartpole_rainbow_config import cartpole_rainbow_config, cartpole_rainbow_create_config
-from .cartpole_iqn_config import cartpole_iqn_config, cartpole_iqn_create_config
-from .cartpole_c51_config import cartpole_c51_config, cartpole_c51_create_config
-from .cartpole_qrdqn_config import cartpole_qrdqn_config, cartpole_qrdqn_create_config
-from .cartpole_sqn_config import cartpole_sqn_config, cartpole_sqn_create_config
-from .cartpole_ppg_config import cartpole_ppg_config, cartpole_ppg_create_config
-from .cartpole_r2d2_config import cartpole_r2d2_config, cartpole_r2d2_create_config
 from .cartpole_acer_config import cartpole_acer_config, cartpole_acer_create_config
+from .cartpole_c51_config import cartpole_c51_config, cartpole_c51_create_config
 from .cartpole_dqfd_config import cartpole_dqfd_config, cartpole_dqfd_create_config
-from .cartpole_sqil_config import cartpole_sqil_config, cartpole_sqil_create_config
-from .cartpole_sql_config import cartpole_sql_config, cartpole_sql_create_config
+from .cartpole_dqn_config import cartpole_dqn_config, cartpole_dqn_create_config
 from .cartpole_dqn_gail_config import cartpole_dqn_gail_config, cartpole_dqn_gail_create_config
 from .cartpole_gcl_config import cartpole_gcl_ppo_onpolicy_config, cartpole_gcl_ppo_onpolicy_create_config
+from .cartpole_impala_config import cartpole_impala_config, cartpole_impala_create_config
+from .cartpole_iqn_config import cartpole_iqn_config, cartpole_iqn_create_config
+from .cartpole_offppo_config import cartpole_ppo_offpolicy_config, cartpole_ppo_offpolicy_create_config
+from .cartpole_ppg_config import cartpole_ppg_config, cartpole_ppg_create_config
+from .cartpole_ppo_config import cartpole_ppo_config, cartpole_ppo_create_config
+from .cartpole_qrdqn_config import cartpole_qrdqn_config, cartpole_qrdqn_create_config
+from .cartpole_r2d2_config import cartpole_r2d2_config, cartpole_r2d2_create_config
+from .cartpole_rainbow_config import cartpole_rainbow_config, cartpole_rainbow_create_config
+from .cartpole_sqil_config import cartpole_sqil_config, cartpole_sqil_create_config
+from .cartpole_sql_config import cartpole_sql_config, cartpole_sql_create_config
+from .cartpole_sqn_config import cartpole_sqn_config, cartpole_sqn_create_config
 from .cartpole_trex_dqn_config import cartpole_trex_dqn_config, cartpole_trex_dqn_create_config
 from .cartpole_trex_offppo_config import cartpole_trex_ppo_offpolicy_config, cartpole_trex_ppo_offpolicy_create_config
 from .cartpole_trex_onppo_config import cartpole_trex_ppo_onpolicy_config, cartpole_trex_ppo_onpolicy_create_config
