+++ dizoo/classic_control/cartpole/config/cartpole_offppo_config.py
@@ -1,7 +1,7 @@
 from easydict import EasyDict
 
 cartpole_ppo_offpolicy_config = dict(
-    exp_name='cartpole_ppo_offpolicy_seed0',
+    exp_name='cartpole_offppo_seed0',
     env=dict(
         collector_env_num=8,
         evaluator_env_num=5,
@@ -51,6 +51,6 @@ cartpole_ppo_offpolicy_create_config = EasyDict(cartpole_ppo_offpolicy_create_co
 create_config = cartpole_ppo_offpolicy_create_config
 
 if __name__ == "__main__":
-    # or you can enter `ding -m serial -c cartpole_ppo_offpolicy_config.py -s 0`
+    # or you can enter `ding -m serial -c cartpole_offppo_config.py -s 0`
     from ding.entry import serial_pipeline
     serial_pipeline((main_config, create_config), seed=0)
